Move kv pkg into server/storage
