+++ server/storage/endpoint/endpoint.go
@@ -16,7 +16,7 @@ package endpoint
 
 import (
 	"github.com/tikv/pd/server/encryptionkm"
-	"github.com/tikv/pd/server/kv"
+	"github.com/tikv/pd/server/storage/kv"
 )
 
 // StorageEndpoint is the base underlying storage endpoint for all other upper
